refactor: remove unused classes and interfaces
